# backend/docs/swagger.yaml
basePath: /api
...
info:
  contact:
    email: nevvjann@gmail.com
    name: shin
  description: This is an auto-generated API Docs.
  license:
    name: MIT
    url: https://github.com/shinjam/simpleChat/LICENSE
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1.0"
paths:
  /v1/token/renew:
    post:
      consumes:
      - application/json
      description: Renew access and refresh tokens.
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: renew access and refresh tokens
      tags:
      - Token
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"